Migration to field usage, to allow instance-based map handling.

+++ services/atlas-maps/atlas.com/maps/data/map/monster/rest.go
@@ -8,10 +8,10 @@ type RestModel struct {
 	Id       uint32 `json:"-"`
 	Template uint32 `json:"template"`
 	MobTime  uint32 `json:"mob_time"`
-	Team     int32  `json:"team"`
+	Team     int8   `json:"team"`
 	CY       int16  `json:"cy"`
 	F        uint32 `json:"f"`
-	FH       uint16 `json:"fh"`
+	FH       int16  `json:"fh"`
 	RX0      int16  `json:"rx0"`
 	RX1      int16  `json:"rx1"`
 	X        int16  `json:"x"`
