+++ services/atlas-maps/atlas.com/maps/data/map/reactor/processor.go
@@ -29,5 +29,5 @@ func NewProcessor(l logrus.FieldLogger, ctx context.Context) Processor {
 }
 
 func (p *ProcessorImpl) InMapProvider(mapId _map.Id) model.Provider[[]Model] {
-	return requests.SliceProvider[RestModel, Model](p.l, p.ctx)(requestReactors(uint32(mapId)), Extract, model.Filters[Model]())
+	return requests.SliceProvider[RestModel, Model](p.l, p.ctx)(requestReactors(mapId), Extract, model.Filters[Model]())
 }
